+++ examples/selection.rs
@@ -9,7 +9,7 @@ use byteorder::{NativeEndian, WriteBytesExt};
 use sctk::data_device::{DataDevice, DndEvent, ReadPipe};
 use sctk::keyboard::{map_keyboard_auto, Event as KbEvent, KeyState};
 use sctk::utils::{DoubleMemPool, MemPool};
-use sctk::window::{BasicFrame, Event as WEvent, Window};
+use sctk::window::{ConceptFrame, Event as WEvent, Window};
 use sctk::Environment;
 
 use sctk::reexports::client::protocol::wl_compositor::RequestsTrait as CompositorRequests;
@@ -45,7 +45,7 @@ fn main() {
     let next_action = Arc::new(Mutex::new(None::<WEvent>));
 
     let waction = next_action.clone();
-    let mut window = Window::<BasicFrame>::init_from_env(&env, surface, dimensions, move |evt| {
+    let mut window = Window::<ConceptFrame>::init_from_env(&env, surface, dimensions, move |evt| {
         let mut next_action = waction.lock().unwrap();
         // Keep last event in priority order : Close > Configure > Refresh
         let replace = match (&evt, &*next_action) {